Fixed notes from mentor
