+++ spec/features/oauth_auth_feature_spec.rb
@@ -5,21 +5,19 @@ RSpec.describe 'Signup with social accounts', type: :feature do
     before { stub_omniauth_provider('github') }
 
     it 'creates user' do
-      expect do
-        visit '/users/sign_in'
-        click_link('Sign in with Github')
-      end.to change(User, :count).by(1)
+      visit '/users/sign_in'
+      click_link('Sign in with Github')
+      page.should have_content('Successfully authenticated from social account.')
     end
   end
 
   context 'when signing up with google' do
     before { stub_omniauth_provider('google_oauth2') }
 
-    it 'visit sign up' do
-      expect do
-        visit '/users/sign_in'
-        click_link('Sign in with Google Oauth2')
-      end.to change(User, :count).by(1)
+    it 'creates user' do
+      visit '/users/sign_in'
+      click_link('Sign in with Google Oauth2')
+      page.should have_content('Successfully authenticated from social account.')
     end
   end
 end
