+++ spec/rails_helper.rb
@@ -22,4 +22,5 @@ RSpec.configure do |config|
   config.include OmniauthHelper
   config.include FeatureTestsHelper
   config.include CapybaraHelper
+  config.include TransactionsFeatureHelper
 end
